Minor fixes to Roundcube routine in upgrade script
Ensure that /var/lib/roundcube exists before attempting to write robots.txt file.

+++ install/upgrade/versions/latest.sh
@@ -114,14 +114,16 @@ if [ -z "$GZIP_LVL_CHECK" ]; then
     $BIN/v-change-sys-config-value "BACKUP_GZIP" '9'
 fi
 
-# Randomize Rouncube des_key for better security
+# Randomize Roundcube des_key for better security
 if [ -f "/etc/roundcube/config.inc.php" ]; then
     rcDesKey="$(openssl rand -base64 30 | tr -d "/" | cut -c1-24)"
     sed -i "s/vtIOjLZo9kffJoqzpSbm5r1r/$rcDesKey/g" /etc/roundcube/config.inc.php
 fi
 
 # Place robots.txt to prevent webmail crawling by search engine bots.
-if [ ! -f "/var/lib/roundcube/robots.txt" ]; then
-    echo "User-agent: *" > /var/lib/roundcube/robots.txt
-    echo "Disallow: /" >> /var/lib/roundcube/robots.txt
+if [ -e "/var/lib/roundcube/" ]; then
+    if [ ! -f "/var/lib/roundcube/robots.txt" ]; then
+        echo "User-agent: *" > /var/lib/roundcube/robots.txt
+        echo "Disallow: /" >> /var/lib/roundcube/robots.txt
+    fi
 fi
